+++ modules/editor/QuickTags.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Enlighter\editor;
+
+use \Enlighter\skltn\ResourceManager as ResourceManager;
+
+class QuickTags{
+
+    public function integrate(){
+        // text editor plugin
+        ResourceManager::enqueueScript('enlighter-quicktags', 'texteditor/QuickTags.js', array('jquery'));
+    }
+}